Add unidirectional named pipe replacements for anonymous pipes
Anonymous pipes in Windows do not support asynchronous I/O, which is
particularly problematic for Go. Consider that you cannot close
an anonymous pipe that is also being Read from another goroutine without
the close blocking. Named pipes, however, fully support asynchronous
operation, but can still be used in many of the same ways anonymous
pipes can be used.

Unfortunately, as of Go 1.10, the stock Go runtime does not use
any of the Windows asynchronous support. So in order to fully exploit
asynchronicity, replacements for anonymous pipe spawning have been
added to this library, in the form of PipeSyncReadAsyncWrite and
PipeAsyncReadSyncWrite.

When these functions are called, the Sync end is intended to be passed
to child processes, while the parent process consumes or produces over
the Async end. The Sync end is returned as an *os.File, so it can be
directly passed to the standard subprocess spawning libraries.

One caveat is that the write ends require the ability to be flushed,
e.g. through os.Sync() (which has since been added for win32File).
There is no standard interface for an io.Writer that also Syncs, so
such an interface has been added as WriteCloserSyncer.