Use atomic.CompareAndSwap to facilitate server-side pipe handle donation
An additional consequence of this change is that win32Pipe.Close()
doesn't _just_ call syscall.CloseHandle, but first calls
disconnectNamedPipe. The nature of the errors returned to clients
consequently changes from a guaranteed io.EOF to syscall.Errno(233).

The client side of the pipe currently does not handle this gracefully.
When it does, consumers of the pipe client relying on syscal.Errno(233)
to be returned will likely not function correctly.