Signal listener close over both internal and external channels
The internal channels are used specifically to signal to win32Pipe.Close
calls that the listener has closed, and will not need donated open
handles in any case. External channels are used to signal to callers of
win32PipeHandler.Accept and win32PipeHandler.Close that the listener is
closed.

This removes a potential race condition with win32PipeListener.Close
not blocking to close an unallocated handle.