Treat ERROR_PIPE_NOT_CONNECTED as io.EOF
This makes sense for net.Conn consumers of general pipe servers: the
pipe server may elect to DisconnectNamedPipe instead of closing the
handle, but it should be treated as io.EOF anyway.

Internally, the only way to trigger an ERROR_PIPE_NOT_CONNECTED is
through the server calling DisconnectNamedPipe. DialPipe calls
CreateFile with OPEN_EXISTING, which makes it impossible for CreateFile
to create a named pipe instance that has not, at some point, been
connected to a server.

The win32PipeListener and win32Pipe implementations now use
DisconnectNamedPipe internally, so this establishes consistency within
the library. Clients using win32Pipe from prior versions will have to be
updated.